Move types into correct package
